chore(categories): updated service method to not allow removal of default categories

+++ src/main/java/com/kunzel/finance_tracker/category/CategoryService.java
@@ -42,6 +42,11 @@ public class CategoryService {
 
   public void removeCategory(Long categoryId) {
     Category categoryToRemove = getCategoryById(categoryId);
+
+    if (categoryToRemove.isDefault()) {
+      throw new IllegalArgumentException("Categoria padrão não pode ser deletada");
+    }
+
     categoryRepository.delete(categoryToRemove);
     // TODO: fazer um cascade delete em TRANSACTIONS com mesmo category_id quando
     // uma categoria for removida
